+++ portainer-monitor-addon/source/Mqtt/HaEntities/HaSensor.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using System.Threading.Tasks;
 
 /// <summary>
 /// Home Assistant Sensor Entity
@@ -105,23 +106,25 @@ internal class HaSensor<T> : HaEntityBase where T : IConvertible
     }
 
     /// <inheritdoc />
-    protected override void OnMqttConnectionStateChanged(object? sender, bool e)
+    protected override Task OnMqttConnectionStateChangedAsync(ConnectionStateChangedEventArgs e)
     {
-        base.OnMqttConnectionStateChanged(sender, e);
-        if (!IsDisposed & e)
+        if (!IsDisposed && e.IsConnected)
         {
             _value = default;
         }
+
+        return base.OnMqttConnectionStateChangedAsync(e);
     }
 
     /// <inheritdoc />
-    protected override void OnHomeAssistantAvailabilityChanged(object? sender, bool e)
+    protected override Task OnHomeAssistantAvailabilityChangedAsync(AvailabilityChangedEventArgs e)
     {
-        base.OnHomeAssistantAvailabilityChanged(sender, e);
-        if (!IsDisposed && e)
+        if (!IsDisposed && e.IsAvailable)
         {
             _value = default;
         }
+
+        return base.OnHomeAssistantAvailabilityChangedAsync(e);
     }
 
     #endregion