+++ portainer-monitor-addon/source/Mqtt/HaEntities/HaUpdate.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using System.Threading.Tasks;
 
 /// <summary>
 /// Home Assistant Update Entity
@@ -129,25 +130,27 @@ internal class HaUpdate : HaEntityBase
     protected override string GetStateMqttPayload() => JsonSerializer.Serialize(_updateState);
 
     /// <inheritdoc />
-    protected override void OnMqttConnectionStateChanged(object? sender, bool e)
+    protected override Task OnMqttConnectionStateChangedAsync(ConnectionStateChangedEventArgs e)
     {
-        base.OnMqttConnectionStateChanged(sender, e);
-        if (!IsDisposed & e)
+        if (!IsDisposed && e.IsConnected)
         {
             _updateState.CurrentVersion = null;
             _updateState.LatestVersion = null;
         }
+
+        return base.OnMqttConnectionStateChangedAsync(e);
     }
 
     /// <inheritdoc />
-    protected override void OnHomeAssistantAvailabilityChanged(object? sender, bool e)
+    protected override Task OnHomeAssistantAvailabilityChangedAsync(AvailabilityChangedEventArgs e)
     {
-        base.OnHomeAssistantAvailabilityChanged(sender, e);
-        if (!IsDisposed && e)
+        if (!IsDisposed && e.IsAvailable)
         {
             _updateState.CurrentVersion = null;
             _updateState.LatestVersion = null;
         }
+
+        return base.OnHomeAssistantAvailabilityChangedAsync(e);
     }
 
     #endregion